Part 6 Exc 6.8 anecdotes, step 6
