+++ part6/redux-anecdotes/src/App.js
@@ -1,30 +1,13 @@
 import React from 'react'
-import { addVote } from './reducers/anecdoteReducer'
 import AnecdoteForm from './components/AnecdoteForm'
+import AnecdoteList from './components/AnecdoteList'
 
 const App = (props) => {
-  const anecdotes = props.store.getState()
-
-  const vote = (id) => {
-    props.store.dispatch(
-      addVote(id)
-    )
-  }
 
   return (
     <div>
       <h2>Anecdotes</h2>
-      {anecdotes.map(anecdote =>
-        <div key={anecdote.id}>
-          <div>
-            {anecdote.content}
-          </div>
-          <div>
-            has {anecdote.votes}
-            <button onClick={() => vote(anecdote.id)}>vote</button>
-          </div>
-        </div>
-      )}
+      <AnecdoteList store={props.store}/>
       <AnecdoteForm store={props.store}/>
     </div>
   )
